pyproject.toml
[project]
name = "paid-edge-backend-api"
version = "0.1.0"
description = "PaidEdge backend API — FastAPI + Supabase + ClickHouse"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115,<1",
    "uvicorn[standard]>=0.34,<1",
    "pydantic-settings>=2.7,<3",
    "supabase>=2.13,<3",
    "clickhouse-connect>=0.8,<1",
    "PyJWT>=2.10,<3",
    "python-multipart>=0.0.18",
    "httpx>=0.28,<1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8,<9",
    "pytest-asyncio>=0.25,<1",
    "httpx>=0.28,<1",
    "ruff>=0.9,<1",
]

[build-system]
requires = ["setuptools>=75"]
build-backend = "setuptools.build_meta"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
